+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/valhalla/ValhallaCurrency.java
@@ -7,15 +7,19 @@ import com.elmakers.mine.bukkit.api.magic.CasterProperties;
 import com.elmakers.mine.bukkit.api.magic.Mage;
 import com.elmakers.mine.bukkit.economy.BaseMagicCurrency;
 
+import me.athlaeos.valhallammo.event.PlayerSkillExperienceGainEvent;
 import me.athlaeos.valhallammo.playerstats.profiles.Profile;
 import me.athlaeos.valhallammo.playerstats.profiles.ProfileCache;
+import me.athlaeos.valhallammo.skills.skills.Skill;
 
-public class ValhallaProfileCurrency extends BaseMagicCurrency {
-    private final Profile profile;
+public class ValhallaCurrency extends BaseMagicCurrency {
+    private final Skill skill;
+    private final Class<? extends Profile> profileClass;
 
-    public ValhallaProfileCurrency(ValhallaManager valhallaManager, Profile profile, String skillId, ConfigurationSection configuration) {
+    public ValhallaCurrency(ValhallaManager valhallaManager, Skill skill, String skillId, ConfigurationSection configuration) {
         super(valhallaManager.getController(), skillId, configuration);
-        this.profile = profile;
+        this.skill = skill;
+        this.profileClass = skill.getProfileType();
     }
 
     @Override
@@ -24,7 +28,7 @@ public class ValhallaProfileCurrency extends BaseMagicCurrency {
         if (player == null) {
             return 0.0;
         }
-        Profile playerProfile = ProfileCache.getOrCache(player, profile.getClass());
+        Profile playerProfile = ProfileCache.getOrCache(player, profileClass);
         if (playerProfile == null) {
             return 0.0;
         }
@@ -38,11 +42,20 @@ public class ValhallaProfileCurrency extends BaseMagicCurrency {
 
     @Override
     public void deduct(Mage mage, CasterProperties caster, double amount) {
-
+        Player player = mage.getPlayer();
+        if (player == null) {
+            return;
+        }
+        Profile playerProfile = ProfileCache.getOrCache(player, profileClass);
+        if (playerProfile == null) {
+            return;
+        }
+        playerProfile.setTotalEXP(playerProfile.getTotalEXP() - amount);
     }
 
     @Override
     public boolean give(Mage mage, CasterProperties caster, double amount) {
-        return false;
+        skill.addEXP(mage.getPlayer(), amount, false, PlayerSkillExperienceGainEvent.ExperienceGainReason.PLUGIN);
+        return true;
     }
 }